renamed vat soap client test
