+++ back/src/companies/vat/vies/client.ts
@@ -3,7 +3,10 @@ import { ASTNode, GraphQLError } from "graphql";
 import { checkVAT } from "jsvat";
 import { createClientAsync, Client, IOptions } from "soap";
 import { CompanyVatSearchResult, ViesResult } from "./types";
-import { countries } from "../../../common/constants/companySearchHelpers";
+import {
+  countries,
+  isVat
+} from "../../../common/constants/companySearchHelpers";
 import logger from "../../../logging/logger";
 import { ErrorCode } from "../../../common/errors";
 
@@ -38,17 +41,16 @@ export const client = async (
     vatNumber,
     countries
   );
+
+  if (!isValid || !isVat(vatNumber)) {
+    throw new UserInputError("Le numéro de TVA n'est pas valide", {
+      invalidArgs: ["clue"]
+    });
+  }
+
   if (!isSupportedCountry) {
     throw new UserInputError(
-      "Le code pays du numéro de TVA intracommunautaire n'est pas valide, veuillez utiliser un code pays ISO à 2 lettres",
-      {
-        invalidArgs: ["clue"]
-      }
-    );
-  }
-  if (!isValid || !vatNumber) {
-    throw new UserInputError(
-      "Le numéro de TVA intracommunautaire n'est pas valide",
+      "Le code pays du numéro de TVA n'est pas valide, veuillez utiliser un code pays intra-communautaire ISO à 2 lettres",
       {
         invalidArgs: ["clue"]
       }
@@ -66,13 +68,14 @@ export const client = async (
     if (viesResult.valid === false) {
       // 404 "no results found"
       throw new UserInputError(
-        "Aucun établissement trouvé avec ce numéro TVA intracommunautaire",
+        "Aucun établissement trouvé avec ce numéro TVA",
         {
           invalidArgs: ["clue"]
         }
       );
     }
 
+    // auto-correct VIES "unknown data"
     const address = viesResult.address === "---" ? "" : viesResult.address;
     const name = viesResult.name === "---" ? "" : viesResult.name;
 
@@ -86,22 +89,29 @@ export const client = async (
       etatAdministratif: "A"
     };
   } catch (err) {
+    // log the error to follow VIES service unavailibility
     logger.error(
-      `Erreur de requete au serveur de recherche de TVA VIES EC ${getReadableErrorMsg(
+      `Erreur with VAT search VIES client: ${getReadableErrorMsg(
         err.root?.Enveloppe?.Body?.Fault?.faultstring
       )}`,
       err
     );
 
     // forward the original exception
-    if (
-      err instanceof UserInputError ||
-      err.root?.Enveloppe?.Body?.Fault?.faultstring === "INVALID_INPUT"
-    ) {
+    if (err instanceof UserInputError) {
       throw err;
     }
 
-    // VIES Server unavailable
+    // throws UserInputError when VIES client returns the error "INVALID_INPUT"
+    if (err.root?.Enveloppe?.Body?.Fault?.faultstring === "INVALID_INPUT") {
+      throw new UserInputError(
+        "Le numéro de TVA recherché n'est pas reconnu par le service de recherche par TVA de la commission européenne (VIES)",
+        {
+          invalidArgs: ["clue"]
+        }
+      );
+    }
+    // Throws VIES Server unavailability message
     if (
       [
         "SERVICE_UNAVAILABLE",
@@ -113,7 +123,7 @@ export const client = async (
       ].includes(err.root?.Enveloppe?.Body?.Fault?.faultstring)
     ) {
       throw new GraphQLError(
-        "Erreur serveur externe de recherche par numéro de TVA de la commission européenne (VIES), veuillez réessayer dans quelques minutes ou si l'erreur persiste, envoyez un mail à contact@trackdechets.beta.gouv.fr avec le contexte de votre erreur",
+        getReadableErrorMsg(err.root?.Enveloppe?.Body?.Fault?.faultstring),
         {
           extensions: {
             code: ErrorCode.EXTERNAL_SERVICE_ERROR
@@ -122,7 +132,7 @@ export const client = async (
       );
     }
 
-    // any VIES Server error
+    // Throws a generic VIES Server error
     throw new GraphQLError(err.message, {
       extensions: {
         code: ErrorCode.EXTERNAL_SERVICE_ERROR
@@ -137,20 +147,20 @@ export const client = async (
 export const getReadableErrorMsg = (faultstring: string): string => {
   switch (faultstring) {
     case "INVALID_INPUT":
-      return "The provided CountryCode is invalid or the VAT number is empty";
+      return "Le code pays ou le numéro de TVA est invalide";
     case "SERVICE_UNAVAILABLE":
-      return "The VIES VAT service is unavailable, please try again later";
+      return "Le service de recherche par TVA de la commission européenne (VIES) est indisponible, veuillez réessayer dans quelques minutes";
     case "MS_UNAVAILABLE":
-      return "The VAT database of the requested member country is unavailable, please try again later";
+      return "Le service de recherche par TVA du pays d'origine est indisponible, veuillez réessayer dans quelques minutes";
     case "MS_MAX_CONCURRENT_REQ":
-      return "The VAT database of the requested member country has had too many requests, please try again later";
+      return "Le service de recherche par TVA du pays d'origine reçoit un trop grand nombre de requêtes, veuillez réessayer dans quelques minutes";
     case "TIMEOUT":
-      return "The request to VAT database of the requested member country has timed out, please try again later";
+      return "Le service de recherche par TVA de la commission européenne (VIES) est indisponible, veuillez réessayer dans quelques minutes";
     case "SERVER_BUSY":
-      return "The service cannot process your request, please try again later";
+      return "Le service de recherche par TVA de la commission européenne (VIES) est indisponible, veuillez réessayer dans quelques minutes";
     case "INVALID_REQUESTER_INFO":
-      return "The requester info is invalid";
+      return "Le service de recherche par TVA de la commission européenne (VIES) est indisponible, veuillez réessayer dans quelques minutes";
     default:
-      return "Unknown error";
+      return "Le service de recherche par TVA de la commission européenne (VIES) est indisponible, veuillez réessayer dans quelques minutes";
   }
 };
